Streamline and improve the content
I have made quite a few changes that aim to improve the guide:

 * Tried to tighten the language
 * Used 'you' in place of 'we' to match the style guide
 * Replaced zip example with one based on `clean`
 * Removed the section on `gradle properties`

The zip example seemed to be more of the same, i.e. just another task that does
something similar to the copy. I think introducing `clean` here answers a
common question, while providing a solid reason to apply the Base Plugin to all
projects.

The output from `gradle properties` doesn't match what was in the guide, so I
removed the section. The command lists a whole bunch of internal stuff that
provide no utility for the user. I don't think this command is particularly
useful at all except for debugging, and even then I tend to print the values of
properties that I'm interested in from within the build.

Signed-off-by: Peter Ledbrook <peter@gradle.com>